Fix paranoid

+++ plugins/paranoid.plugin.zsh
@@ -7,19 +7,15 @@ function burry() {
         echo 'Error: gpg is not installed.' >&2
         exit 1
     fi
-
     if [ $# -eq 0 ]; then
-        echo "No arguments supplied"
-    fi
-
-    if [ -z "${$1}" ]; then
-        echo "No arguments supplied"
+       echo "No folder supplied"
+       return
     fi
 
     tar -cz $1 | gpg -e -o $1.crypt
-    cp $1.crypt $1.crypt.bak
 
-    if [ -f $1.crypt ]; then
+    sta=$?
+    if [ $sta -eq 0 ] && [ -f $1.crypt ]; then
         find $1 -type f -exec shred -u {} \;
         rm -rf $1
     fi
